requirements.txt
waitress>=2.1.2
flask>=2.3.3
flask-httpauth>=4.8.0
Flask_RESTful>=0.3.10
python-dotenv>=1.0.0
openai>=1.31.2
etcd3
protobuf==3.20.3